refactor(web): extract gateway-topology readout into web_gateway_topology (web_server.py god-file slice R2-B)
